Move link/unlink under context with top-level aliases
